+++ src/library/DIPS.Mobile.UI/Components/ListItems/ListItem.Properties.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel;
 using System.Windows.Input;
+using DIPS.Mobile.UI.Components.ContextMenus;
 using DIPS.Mobile.UI.Components.ListItems.Options;
 using DIPS.Mobile.UI.Components.ListItems.Options.ContextMenu;
 using DIPS.Mobile.UI.Components.ListItems.Options.Dividers;
@@ -177,6 +178,18 @@ namespace DIPS.Mobile.UI.Components.ListItems
             set => SetValue(PaddingProperty, value);
         }
 
+        public ContextMenu ContextMenu
+        {
+            get => (ContextMenu)GetValue(ContextMenuProperty);
+            set => SetValue(ContextMenuProperty, value);
+        }
+        
+        public static readonly BindableProperty ContextMenuProperty = BindableProperty.Create(
+            nameof(ContextMenu),
+            typeof(ContextMenu),
+            typeof(ListItem),
+            propertyChanged: (bindable, _, _) => ((ListItem)bindable).AddContextMenu());
+        
         public new static readonly BindableProperty MarginProperty = BindableProperty.Create(
             nameof(Margin),
             typeof(Thickness),
@@ -214,6 +227,7 @@ namespace DIPS.Mobile.UI.Components.ListItems
             nameof(Command),
             typeof(ICommand),
             typeof(ListItem),
+            defaultValue:null,
             propertyChanged: (bindable, _, _) => ((ListItem)bindable).SetTouchIsEnabled());
     
         public static new readonly BindableProperty BackgroundColorProperty = BindableProperty.Create(
@@ -292,7 +306,8 @@ namespace DIPS.Mobile.UI.Components.ListItems
             nameof(ContextMenuOptions),
             typeof(ContextMenuOptions),
             typeof(ListItem),
-            propertyChanged: (bindable, _, _) => ((ListItem)bindable).AddContextMenu());
+            defaultValueCreator: CreateOptionsAndBind<ContextMenuOptions>,
+            propertyChanged: (bindable, _, newValue) => ((ContextMenuOptions)newValue).Bind((ListItem)bindable));
     
         private static T CreateOptionsAndBind<T>(BindableObject bindable) where T : ListItemOptions, new()
         {