+++ src/library/DIPS.Mobile.UI/Components/ListItems/Options/ContextMenu/ContextMenuOptions.Properties.cs
@@ -4,25 +4,21 @@ namespace DIPS.Mobile.UI.Components.ListItems.Options.ContextMenu;
 
 public partial class ContextMenuOptions
 {
-    public static readonly BindableProperty MenuProperty = BindableProperty.Create(
-        nameof(Menu),
-        typeof(ContextMenus.ContextMenu),
-        typeof(ContextMenuOptions));
-
+   
     public static readonly BindableProperty ModeProperty = BindableProperty.Create(
         nameof(Mode),
         typeof(ContextMenuEffect.ContextMenuMode),
-        typeof(ContextMenuOptions));
+        typeof(ContextMenuOptions),
+        defaultValue: ContextMenuEffect.ContextMenuMode.LongPressed);
 
+    /// <summary>
+    /// The mode of the Context Menu
+    /// <remarks>Default is 'LongPressed'</remarks>
+    /// </summary>
     public ContextMenuEffect.ContextMenuMode Mode
     {
         get => (ContextMenuEffect.ContextMenuMode)GetValue(ModeProperty);
         set => SetValue(ModeProperty, value);
     }
     
-    public ContextMenus.ContextMenu Menu
-    {
-        get => (ContextMenus.ContextMenu)GetValue(MenuProperty);
-        set => SetValue(MenuProperty, value);
-    }
 }