+++ src/library/DIPS.Mobile.UI/Components/ListItems/Options/ContextMenu/ContextMenuOptions.cs
@@ -7,6 +7,5 @@ public partial class ContextMenuOptions : ListItemOptions
     public override void DoBind(ListItem listItem)
     {
         ContextMenuEffect.SetMode(listItem.Border, Mode);
-        ContextMenuEffect.SetMenu(listItem.Border, Menu);
     }
 }